fix: flv live context undefined

+++ packages/xgplayer-flv-live/src/mse/index.js
@@ -183,6 +183,7 @@ class FlvPlayer extends BasePlugin {
   }
 
   _destroy () {
+    if (!this.context) return Promise.resolve()
     return this.flv.mse.destroy().then(() => {
       this.context.destroy()
       this.flv = null
